SoldItems READY for deployment

+++ BackEnd/DishServices/demo/src/main/java/com/example/demo/model/Dish.java
@@ -14,25 +14,23 @@ import lombok.NoArgsConstructor;
 @Entity
 @Table(name = "dish")
 public class Dish {
-@Id
-@GeneratedValue(strategy = GenerationType.AUTO)
-@Column(name ="id" , nullable = false)
-private int id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id", nullable = false)
+    private int id;
 
-@Column(name = "dish_name", nullable = false)
-private String dish_name;
+    @Column(name = "dish_name", nullable = false)
+    private String dish_name;
 
-@Column(name="price", nullable = false)
-private Double price;
+    @Column(name = "price", nullable = false)
+    private Double price;
 
-@Column(name ="amount", nullable = false)
-private Integer amount;
+    @Column(name = "amount", nullable = false)
+    private Integer amount;
 
-@Column(name = "sale", nullable = false)
-private boolean sale = false;
+    @Column(name = "sale", nullable = false)
+    private boolean sale = false;
 
-@Column (name = "company_id", nullable = false)
-private int company_id;
-
+    @Column(name = "company_id", nullable = false)
+    private Long company_id;
 }
-
